+++ .gitignore
@@ -2,3 +2,4 @@ cache.json
 packages/
 test_output
 test_env/
+venv/
